Sketch the structure of the improved main program
Doing the branch coverage should not prevent the programmers from
continuing to work on the original files. Plus, it is always dangerous
to edit the source code in its original location. Therefore all the code
to be instrumented will be copied to a temporary directory first and
then instrumented there.