Rework command line parsing
Before, the command line had a non-standard format, which was not
necessary. The only option that has been passed through to "go test" was
"-vet=off", and this is better handled by a normal command line option
instead of affecting the command line syntax for the normal case.

Also added -help and -version options.

Fixes #5