Allow the coverage counts to be written to disk immediately
When the coverage data is collected in separate runs of the target
program, some of these program runs may use os.Exit and thus prevent
any deferred functions from being run. For these cases, the -immediately
option writes the coverage counts back to disk whenever a coverage point
is visited. This makes the coverage much slower because of the I/O
overhead, but it should only be necessary in a few cases anyway.

If this option is really required, it would probably be more efficient
to just append the (idx, cond) data to a temporary file instead of
generating the whole JSON data again and again. The JSON data is about
400 kB in size for github.com/rillig/pkglint, with less than 7000
conditions. In a small ad-hoc benchmark, writing the JSON data to disk
took 5 milliseconds, and several of the conditions are visited more than
a million times.

See https://github.com/rillig/gobco/issues/2