Allow the coverage data to be filled across multiple sessions
When running `gobco -test -c -keep`, the instrumented test program is
created and placed into the temporary directory. Because of the `-keep`,
the temporary directory is not removed afterwards.

Since gobco expects `go test` to always create the file containing the
coverage data, it currently quits, without even saying where the
temporary directory is. But it's there.

The instrumented test program in the temporary directory now requires
the GOBCO_STATS environment variable to be set. When this is done, it
collects the coverage data into that file.

See https://github.com/rillig/gobco/issues/2