Release gobco 0.9.4
Changes since 0.9.3:

Switch statements with initialization statements are instrumented in
several (but not all) cases.

Generated code (such as from stringer or goyacc) is not instrumented.

The code that does the instrumentation is embedded in strings, and these
strings are included in the distribution. It is no longer necessary to
run "go generate" after downloading gobco.