Clean up README
Starting the introduction with the use case of the failing tests does
not make a good first impression since the example output is a wild
mixture from several programs, and there is only a single line of
interesting output.  Starting with the pkglint example provides a more
realistic impression of gobco's capabilities.

Reword the section about inserting custom conditions in the source code
since the previous version was not as clear as possible.