Prepare supporting Go modules
In order to clearly express how a single argument is processed, almost
the whole code had to be rewritten. The main change is the new type
argInfo that describes how an argument is processed in the later phases.
This part will differ between Go modules and traditional Go packages.

No functional change.