For Go modules, instrument the source directory as well
Previously, the module root directory in the temporary build environment
was instrumented, which was completely wrong since in the test case
'testdata/test_main', the gobco code itself was instrumented, not the
test_main code.

The tests had passed before since they did not exercise the code for Go
modules. This is because locally I had tested the code from its
traditional location, '$HOME/go/src/github.com/rillig/gobco'. Having two
copies of the code checked out was definitely confusing. The Travis
tests ran in that same configuration since Travis defaults to using
GO111MODULES=auto.

Not sure whether the change to .travis.yml is correct, as Travis
currently "Could not authorize build request for rillig/gobco", but it
doesn't tell me why. That's also the reason why I didn't get any build
reports for the previous 7 commits. The last that succeeded was the one
for release 0.10.0.