Fix instrumenting of non-modules
For non-modules, the wrong directory was passed to the instrumenter,
resulting in a panic because of a nonexistent file.

In installations that have GO111MODULE set, this variable interfered
with 'go test', leading to the error message 'go: go.mod file not found
in current directory or any parent directory'.

Fixes https://github.com/rillig/gobco/issues/20.