Fix return type of instrumenter.nextVarname
By returning only the name of the variable instead of an Indent node,
the calling code is forced to create new identifier nodes as needed, so
that the AST is now a real tree, instead of a directed acyclic graph.

Reusing the identifier nodes in the AST had worked but was probably
undefined behavior.