Fix comment placement when wrapping conditions with gobcoCover
By setting the position of all tokens to the fixed value of the original
condition, any comment with a later position will still be considered to
occur later than the whole new condition.

Previously, the position of the tokens was token.NoPos, which triggered
the automatic position calculation from printer.intersperseComments.

This is the same trick that gofmt uses as well.

The same fix needs to be applied to all other newly created AST nodes as
well, when instrumenting switch statements. This will be done in a
follow-up commit.