Fix placement of comments in instrumented code, mostly
In type switch statements, the type names from the case clauses are used
earlier in the instrumented code than in the original code. This leads
to wrong placement of comments, for example in TypeSwitchStmt.gobco:114.

To fix this, it may be necessary to use reflection once more, to find
the first struct field. On the other hand, go/ast only defines 5 type
constructors, plus identifiers, so a brute force approach may work as
well.