Fix instrumentation of types based on bool
The predefined type `bool` is not interchangeable with custom-defined
boolean types.

The result of `==` is a `bool` type compatible with both the predefined
`bool` and derived boolean types, while `!` preserves the derived type,
so use the `==` operator to convert between the boolean types. While
this adds some bloat to the instrumented code, it still allows gobco to
work without resolving the actual types. Working with the actual types
may be more elegant but is left for later.

Fixes https://github.com/rillig/gobco/issues/35.